+++ documentation/RELEASE_NOTES.md
@@ -1,3 +1,18 @@
+EPICS 7 release 4.7.1
+=====================
+
+API changes to PvaClientMonitor
+-------------------------------
+
+The create method that had arguments for stateChangeRequester and monitorRequester no longer exists.
+
+API changes to PvaClientGet, ..., PvaClientMonitor
+--------------------------------------------------
+
+pvaClientGet, ..., pvaClientMonitor all implemented PvaClientChannelStateChangeRequester.
+This was never called and has been removed.
+
+
 EPICS V4 release 4.7
 ====================